requirements.txt
astroid==3.3.9
coverage==7.6.12
dill==0.3.9
iniconfig==2.0.0
isort==6.0.1
mccabe==0.7.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.5
pytest==8.3.5
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tomlkit==0.13.2
tzdata==2025.1
